Ugly bug fix for red light not showing when index = 0 in Versus
Bug description:
On initial start, if the randomized target is 0 (target = 0), red light won't be assigned.
The colour member in Block is not set;
the actual led colour CANNOT BE SET either (you can, but it won't work).
I have not fully tested this but it seems like the first and the last LEDs cannot be set. If so, it could be related to this issue (https://github.com/FastLED/FastLED/issues/583).

This could be caused by the conflict between FastLED and I2C libraries.

Current "fix": Forcefully assign target = 1 when initial target = 0 for red.